utilize task hydration to improve performance
